add correction partie 1

+++ nodefs.js
@@ -2,42 +2,48 @@ const fs = require('fs')
 
 //lire un fichier dans ce cas là créer un fichier avec le texte precis
 if(process.argv[2] === 'echo'){
-switch (process.argv.length) {
-    case 2:
-        console.log(`usage echo : node nodefs.js echo 'nom du fichier a créer' 'texte a mettre dans le fichier'\n`)
-        process.exit(1)
-    case 3:
-        console.log(`usage echo : node nodefs.js echo 'nom du fichier a créer' 'texte a mettre dans le fichier'\n`)
-        process.exit(1)
-    default:
+//verifier la cmd line
+if(process.argv.length !== 4){
+    console.log('usage: node echo file text')
+    process.exit(1)
+}
+
+//check if file exist
+if(!fs.existsSync(process.argv[2])){
+    console.log(`Error: ${process.argv[2]} does not exist`)
+    process.exit(1)
 }
 fs.writeFileSync(process.argv[3], process.argv[4])
 }
 
 // fait une copie du fichier a dans le nouveau fichier B
 if(process.argv[2] === 'cp'){
-    switch (process.argv.length) {
-        case 2:
-            console.log(`usage echo : node nodefs.js cp 'nom du fichier a copier' 'nom du nouveau fichier qui contiendra la copie'\n`)
-            process.exit(1)
-        case 3:
-            console.log(`usage echo : node nodefs.js cp 'nom du fichier a copier' 'nom du nouveau fichier qui contiendra la copie'\n`)
-            process.exit(1)
-        default:
-    }
+ //verifier la cmd line
+if(process.argv.length !== 4){
+    console.log('usage: node cp fileA fileB')
+    process.exit(1)
+}
+
+//check if file exist
+if(!fs.existsSync(process.argv[2])){
+    console.log(`Error: ${process.argv[2]} does not exist`)
+    process.exit(1)
+}
 fs.copyFileSync(process.argv[3], process.argv[4])
 }
 
 //créer un nouveau dossier
 if(process.argv[2] === 'mkdir'){
-    switch (process.argv.length) {
-        case 2:
-            console.log(`usage echo : node nodefs.js mkdir 'nom du fichier a créer'\n`)
-            process.exit(1)
-        case 3:
-            console.log(`usage echo : node nodefs.js mkdir 'nom du fichier a créer'\n`)
-            process.exit(1)
-        default:
-    }
+//verifier la cmd line
+if(process.argv.length !== 3){
+    console.log('usage: node mkdir')
+    process.exit(1)
+}
+
+//check if file exist
+if(!fs.existsSync(process.argv[2])){
+    console.log(`Error: ${process.argv[2]} does not exist`)
+    process.exit(1)
+}
 mkdirSync(process.argv[3])
 }
